pyproject.toml
[tool.poetry]
name = "diploma-project"
version = "0.1.0"
description = ""
authors = ["asya474 <asya474@gmail.com>"]
readme = "README.md"
packages = [{include = "diploma-project"}]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "8.0.0rc2"
pluggy = "^1.2.0"
requests = "^2.0.0"
python-dotenv = "^1.0.0"
selene = "2.0.0rc6"
selenium = "^4.17.2"
allure-pytest = "^2.9.45"
jsonschema = "^4.4.0"
appium-python-client = "3.1.1"

[tool.poetry.group.dev.dependencies]
pylint = "^3.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--alluredir=allure-results -s --clean-alluredir"